refactor:add fixes to Battle_ground project of 'OOP Principles'
