+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Enemy.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_groung;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
 
 public abstract class Enemy implements Mortal{
     private int health;
