+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Ghost.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_groung;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
 
 public class Ghost extends Witch {
     protected final int DAMAGE_GHOST = 20;
