+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Hero.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_groung;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
 
 public abstract class Hero implements Mortal {
     private String name;
