+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Mortal.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_groung;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
 
 public interface Mortal {
     boolean isAlive();
