+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/battle_ground/Vampire.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.battle_groung;
+package com.rakovets.course.java.core.practice.oop_principles.battle_ground;
 
 public class Vampire extends Enemy{
     private final int DAMAGE_VAMPIRE = 25;
